More delete/rm/unstage cleanup
